pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.0",
    "python-telegram-bot==20.3",
]
